Changing db

+++ src/main/java/com/rog/webshop/dao/order/OrderDao.java
@@ -3,5 +3,5 @@ package com.rog.webshop.dao.order;
 import com.rog.webshop.model.order.Order;
 
 public interface OrderDao {
-    Long saveOrder(Order order);
+    int saveOrder(Order order);
 }
