+++ src/main/java/com/rog/webshop/service/order/OrderService.java
@@ -5,5 +5,5 @@ import com.rog.webshop.model.order.Order;
 public interface OrderService {
     void processOrder(String productId, long quantity);
 
-    Long saveOrder(Order order);
+    int saveOrder(Order order);
 }
